# web/openapi.json
{
  "components": {
    "schemas": {
...
      "Caveat": {
        "description": "A data caveat inserted into the report (PRD \u00a79 step 2).",
        "properties": {
          "kind": {
            "enum": [
              "limited_group",
              "small_sample",
              "proxy_metric",
              "unadjusted_defence",
              "stale_value",
              "no_market_value",
              "no_previous_season",
              "value_model",
              "no_comparables",
              "same_club"
            ],
            "title": "Kind",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "text"
        ],
        "title": "Caveat",
        "type": "object"
      },
...
      "ComparableFact": {
        "description": "A similar player in the same position group (PRD \u00a78.10 step 2).",
        "properties": {
          "player_id": {
            "title": "Player Id",
            "type": "integer"
          },
          "player_name": {
            "title": "Player Name",
            "type": "string"
          },
          "similarity": {
            "description": "Cosine similarity, -1 to 1.",
            "title": "Similarity",
            "type": "number"
          },
          "team_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Name"
          }
        },
        "required": [
          "player_id",
          "player_name",
          "team_name",
          "similarity"
        ],
        "title": "ComparableFact",
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "Machine-readable error details.",
        "properties": {
          "code": {
            "description": "Stable error code, e.g. 'not_found'.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "The single error schema returned by every endpoint.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "FactSheet": {
        "description": "Everything a report about one player may state (PRD \u00a79).",
        "properties": {
          "age": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "as_of": {
            "description": "Date ages and contract windows are measured from.",
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "birth_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "caveats": {
            "items": {
              "$ref": "#/components/schemas/Caveat"
            },
            "title": "Caveats",
            "type": "array"
          },
          "chance_of_playing": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chance Of Playing"
          },
          "comparables": {
            "items": {
              "$ref": "#/components/schemas/ComparableFact"
            },
            "title": "Comparables",
            "type": "array"
          },
          "concerns": {
            "description": "KPI ids rated weak.",
            "items": {
              "type": "string"
            },
            "title": "Concerns",
            "type": "array"
          },
          "contract_expiry": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Expiry"
          },
          "current_season": {
            "title": "Current Season",
            "type": "string"
          },
          "effective_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Blended minutes behind the rates.",
            "title": "Effective Minutes"
          },
          "fit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FitFact"
              },
              {
                "type": "null"
              }
            ]
          },
          "fpl_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fpl Status"
          },
          "implied_value": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImpliedValueFact"
              },
              {
                "type": "null"
              }
            ]
          },
          "kpis": {
            "items": {
              "$ref": "#/components/schemas/KpiFact"
            },
            "title": "Kpis",
            "type": "array"
          },
          "market_value": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MarketValueFact"
              },
              {
                "type": "null"
              }
            ]
          },
          "minutes": {
            "description": "This season's FPL minutes.",
            "title": "Minutes",
            "type": "number"
          },
          "player_id": {
            "title": "Player Id",
            "type": "integer"
          },
          "player_name": {
            "title": "Player Name",
            "type": "string"
          },
          "position_group": {
            "title": "Position Group",
            "type": "string"
          },
          "previous_season": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Set when last season is blended in.",
            "title": "Previous Season"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoleFact"
              },
              {
                "type": "null"
              }
            ]
          },
          "season_mode": {
            "title": "Season Mode",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Receipt"
            },
            "title": "Sources",
            "type": "array"
          },
          "status_as_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status As Of"
          },
          "strengths": {
            "description": "KPI ids rated strong or elite.",
            "items": {
              "type": "string"
            },
            "title": "Strengths",
            "type": "array"
          },
          "team_id": {
            "title": "Team Id",
            "type": "integer"
          },
          "team_name": {
            "title": "Team Name",
            "type": "string"
          }
        },
        "required": [
          "player_id",
          "player_name",
          "team_id",
          "team_name",
          "position_group",
          "birth_date",
          "age",
          "as_of",
          "season_mode",
          "current_season",
          "previous_season",
          "minutes",
          "effective_minutes",
          "fpl_status",
          "chance_of_playing",
          "status_as_of",
          "contract_expiry",
          "market_value",
          "kpis",
          "strengths",
          "concerns",
          "role",
          "comparables",
          "implied_value",
          "fit",
          "caveats",
          "sources"
        ],
        "title": "FactSheet",
        "type": "object"
      },
...
      "MarketValueFact": {
        "description": "A Transfermarkt estimated market value with Transfermarkt's own as-of date.",
        "properties": {
          "is_stale": {
            "title": "Is Stale",
            "type": "boolean"
          },
          "source": {
            "description": "override, transfermarkt or transfermarkt_datasets",
            "title": "Source",
            "type": "string"
          },
          "tm_last_updated": {
            "format": "date",
            "title": "Tm Last Updated",
            "type": "string"
          },
          "value_eur": {
            "title": "Value Eur",
            "type": "integer"
          }
        },
        "required": [
          "value_eur",
          "tm_last_updated",
          "source",
          "is_stale"
        ],
        "title": "MarketValueFact",
        "type": "object"
      },
...
      "PlayerAgeCurveResponse": {
        "description": "Next-season projections for one player from the age curves (US-18).",
        "properties": {
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "caveat": {
            "title": "Caveat",
            "type": "string"
          },
          "effective_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Minutes"
          },
          "player_id": {
            "title": "Player Id",
            "type": "integer"
          },
          "projections": {
            "items": {
              "$ref": "#/components/schemas/ProjectionOut"
            },
            "title": "Projections",
            "type": "array"
          },
          "seasons": {
            "items": {
              "type": "string"
            },
            "title": "Seasons",
            "type": "array"
          }
        },
        "required": [
          "player_id",
          "age",
          "effective_minutes",
          "seasons",
          "caveat",
          "projections"
        ],
        "title": "PlayerAgeCurveResponse",
        "type": "object"
      },
      "PlayerSearchHit": {
        "description": "A player search result.",
        "properties": {
          "kind": {
            "enum": [
              "exact",
              "prefix",
              "fuzzy"
            ],
            "title": "Kind",
            "type": "string"
          },
          "matched": {
            "title": "Matched",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "player_id": {
            "title": "Player Id",
            "type": "integer"
          },
          "position_group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Group"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "team_id": {
            "title": "Team Id",
            "type": "integer"
          },
          "team_name": {
            "title": "Team Name",
            "type": "string"
          }
        },
        "required": [
          "player_id",
          "name",
          "team_id",
          "team_name",
          "position_group",
          "matched",
          "kind",
          "score"
        ],
        "title": "PlayerSearchHit",
        "type": "object"
      },
...
      "ReportResponse": {
        "description": "A grounded scouting report with the fact sheet behind it (US-06).",
        "properties": {
          "facts": {
            "$ref": "#/components/schemas/FactSheet"
          },
          "report": {
            "$ref": "#/components/schemas/ReportOut"
          }
        },
        "required": [
          "report",
          "facts"
        ],
        "title": "ReportResponse",
        "type": "object"
      },
...
      "SimilarResponse": {
        "description": "Top-k similar players (US-09).",
        "properties": {
          "player_id": {
            "title": "Player Id",
            "type": "integer"
          },
          "player_name": {
            "title": "Player Name",
            "type": "string"
          },
          "position_group": {
            "title": "Position Group",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SimilarPlayerOut"
            },
            "title": "Results",
            "type": "array"
          },
          "season_mode": {
            "enum": [
              "blended",
              "current"
            ],
            "title": "Season Mode",
            "type": "string"
          }
        },
        "required": [
          "player_id",
          "player_name",
          "position_group",
          "season_mode",
          "results"
        ],
        "title": "SimilarResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Read-only API over the local warehouse. No number without a receipt.",
    "title": "PL Scout Engine API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/players/search": {
      "get": {
        "description": "Prefix search over this season's players (any word of the name), fuzzy for typos.",
        "operationId": "search_players_players_search_get",
        "parameters": [
          {
            "description": "Player name.",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "Player name.",
              "maxLength": 100,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlayerSearchHit"
                  },
                  "title": "Response Search Players Players Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown club, player or need."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request parameters."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The warehouse has not been built yet."
          }
        },
        "summary": "Search Players",
        "tags": [
          "players"
        ]
      }
    },
    "/players/{player_id}": {
      "get": {
        "description": "Player profile: header facts, percentiles vs position peers, value, role, caveats.\n\nEvery KPI carries its source and as-of; the Transfermarkt estimated market value\ncarries Transfermarkt's own date.",
        "operationId": "player_profile_players__player_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "title": "Player Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "season_mode",
            "required": false,
            "schema": {
              "default": "blended",
              "enum": [
                "blended",
                "current"
              ],
              "title": "Season Mode",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactSheet"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown club, player or need."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request parameters."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The warehouse has not been built yet."
          }
        },
        "summary": "Player Profile",
        "tags": [
          "players"
        ]
      }
    },
    "/players/{player_id}/age-curve": {
      "get": {
        "description": "Next-season projection from the age curves: blended rate + typical change at age.",
        "operationId": "player_age_curve_players__player_id__age_curve_get",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "title": "Player Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerAgeCurveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown club, player or need."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request parameters."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The warehouse has not been built yet."
          }
        },
        "summary": "Player Age Curve",
        "tags": [
          "players"
        ]
      }
    },
    "/players/{player_id}/report": {
      "get": {
        "description": "Grounded scouting report (template, or a validated local-LLM rewrite) and its facts.",
        "operationId": "player_report_players__player_id__report_get",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "title": "Player Id",
              "type": "integer"
            }
          },
          {
            "description": "Club whose need the report addresses.",
            "in": "query",
            "name": "team_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Club whose need the report addresses.",
              "title": "Team Id"
            }
          },
          {
            "in": "query",
            "name": "season_mode",
            "required": false,
            "schema": {
              "default": "blended",
              "enum": [
                "blended",
                "current"
              ],
              "title": "Season Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "benchmark",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "top6",
                    "top4",
                    "league",
                    "custom"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Benchmark"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown club, player or need."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request parameters."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The warehouse has not been built yet."
          }
        },
        "summary": "Player Report",
        "tags": [
          "players"
        ]
      }
    },
    "/players/{player_id}/similar": {
      "get": {
        "description": "Top-k most similar players in the same position group (cosine, US-09).",
        "operationId": "similar_players_players__player_id__similar_get",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "title": "Player Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "k",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "K"
            }
          },
          {
            "in": "query",
            "name": "season_mode",
            "required": false,
            "schema": {
              "default": "blended",
              "enum": [
                "blended",
                "current"
              ],
              "title": "Season Mode",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimilarResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown club, player or need."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request parameters."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The warehouse has not been built yet."
          }
        },
        "summary": "Similar Players",
        "tags": [
          "players"
        ]
      }
    },
...
  }
}